more fixes to alessandro's 'udot' change to grackle
The grackle modifications are a bit of a hot mess.  I reverted a
previous change so that standard grackle functions take a
`gr_float` of `udot`, rather than a `gr_float *`.  I had previously
modified it so that it took `gr_float *` to match the grackle internal
functions (e.g. `_solve_chemistry` and `solve_chemistry_`.

Also, previously this only applied the `udot` change to non-tabulated
grackle functions.  I've now extended this change so that it also
works for the tabulated grackle functions, to keep more symmetry
between the tabulated and non-tabulated versions.

In general, I think the lesson is not to trust any further modifications
to the grackle api without fairly significant testing.  The current
status is already a hot mess.