Improved feedback injection to model particle merging
Problem: the amount of energy that actually gets injected depends
on the timestep during injection. This is bad, as it indicates
something isn't actually conservative.

Fixing this is complicated, because the code doesn't directly
track conservatives (besides mass) and cells aren't updated
at the same code-time.

I hoped things would be better if I more-correctly updated the
primitive predictor fields. So I modelled my feedback mechanism
after the particle merging mechanism.  This is a little difficult,
because the ejecta isn't actually another particle, but I was able to
guess how to adapt most of the fields.

Things still aren't fixed. The timesteps are still set *before* the
thermal energy is injected, and if that timestep is too large,
too much energy will be injected.  I think this has improved
things a bit, but it's not perfect.

Maybe I need to better understand all of the particles fields?
Maybe I should figure out how to reset the timestep when SNe occur?
Maybe I should recompute the primitives?