Inject SNe using python
Previously I was trying to add SNe directly within the C code of gizmo.
This didn't work terribly great.  Even when I was just adding mass and
energy to existing cells, I had some trouble with non-conservative effects
due to this discontinuity (in time) of mass and energy.  Additionally,
as the hot bubble became more and more diffuse, it became more and more
likely that there would be no particles near the SN injection site,
or that the code would crash when you dumped all the energy into 1
particle.

So I decided to switch to injecting new particles.  After looking around
for a bit, it became clear that this was going to be very difficult to do
within gizmo -- this sudden change in the domain broke a number of
assumptions going into the primary gizmo machinery.  So instead I took
a somewhat simpler route: end a simulation just before a SNe, then
create a new snapshot just *after* the SNe and add new particles
to that file.  Then you can restart gizmo with this post-SN snapshot,
and it'll happily evolve it until the next SN.  This has some downsides:
1) It means you have to manually restart the simulation every SN.
2) It looks a bit of time-accuracy, due to the interpolation needed to
synchronize the computational domain at the snapshot time. (I don't think
this is a significant issue. The sudden injection of mass and energy
will also have this problem, even if you did it within gizmo, *I think*.)

So let's see how this does.