add support for doing runs on raijin
Mainly this consists of adding a new type of qsub script, which is
demonstrated in `runs/cluster_cooling_200/inputs/cluster_cooling_200_loop-raijin`.
But that's not the only change. `scripts/prepare_for_restart.py` also was trying
to import a bunch of libraries it didn't actually need (like pygrackle).
So I refactored some of that code, to avoid trying to import unnecessary packages.